feat(decomp): byte-match exe/slus_004_22@0x8014F704

+++ src/exe/slus_004_22/symbols/functions.h
@@ -36,7 +36,7 @@ void func_8015D044(void);
 void func_8014B87C(u16 countdown);
 void func_8014E5A0(u32 ot_index, u32 primitive_size);
 void func_8014F514(s32 a0, s32 a1, s32 a2);
-u8   func_8014F704(s16* value, s32 arg, u8 arg2, u8 arg3, s32 arg4);
+u8   func_8014F704(s16* value, s32 arg, u8 arg2, u8 arg3, u8 arg4);
 void func_80150098(s16 x, s16 y, u32 clut, const u8* text);
 void func_8015DF18(u16 cue_id);
 s32  func_801655F4(u8* counter, s32 delta);
